+++ recipes/_unit_testing_agent_capability.rb
@@ -7,12 +7,12 @@ bamboo_agent_capability 'system.of.a.down' do
 end
 
 template 'bamboo-capabilities.properties' do
-  path "#{node[:bamboo][:agent][:data_dir]}/bin/bamboo-capabilities.properties"
+  path "#{node['bamboo']['agent']['data_dir']}/bin/bamboo-capabilities.properties"
   source 'bamboo-capabilities.properties.erb'
-  owner  node[:bamboo][:agent][:user]
-  group  node[:bamboo][:agent][:group]
+  owner  node['bamboo']['agent']['user']
+  group  node['bamboo']['agent']['group']
   mode '0644'
   variables(
-    :options => node[:bamboo][:agent_capabilities]
+    :options => node['bamboo']['agent_capabilities']
   )
 end